Move logic to PackProfile
