+++ launcher/ui/pages/instance/VersionPage.h
@@ -104,8 +104,6 @@ private:
     void updateButtons(int row = -1);
     void preselect(int row = 0);
     int doUpdate();
-    void fixModLoaderVersions(const QString& gameVersion);
-    void fixModLoaderVersion(const QString& id, const QString& gameVersion);
 
 protected:
     QMenu * createPopupMenu() override;
